Fix: 중복된 상품 추가시 카운트가 올라가지 않는 이슈
