+++ src/app/(provider)/(root)/cart/page.tsx
@@ -5,7 +5,6 @@ import { useAuth } from '@/contexts/auth.context/auth.context';
 import { useCartsMutation } from '@/hooks/mutation';
 import { useCartsQuery } from '@/hooks/query';
 import useLocalCart from '@/hooks/useLocalCart';
-import { CartItem } from '@/types/cart';
 import { useCallback, useEffect, useMemo } from 'react';
 import { CartAccordion, CartEmpty, CartList, CartSelector } from './_components';
 
@@ -91,30 +90,6 @@ const CartPage = () => {
     patchMutation,
     updatedCartItem
   ]);
-
-  useEffect(() => {
-    if (cartList?.length) {
-      const prevLocalCartList = localStorage.getItem('cart');
-
-      const parsedLocalCartList = prevLocalCartList ? JSON.parse(prevLocalCartList) : [];
-
-      const filteredCartList = cartList.filter(
-        (cartItem) =>
-          !parsedLocalCartList.find((localCartItem: CartItem) => localCartItem.productId === cartItem.productId)
-      );
-
-      const newCartList = [...parsedLocalCartList, ...filteredCartList];
-
-      localStorage.setItem('cart', JSON.stringify(newCartList));
-      setLocalCartList(newCartList);
-    } else {
-      const prevCartList = localStorage.getItem('cart');
-      if (prevCartList) {
-        const parsedPrevCartList = JSON.parse(prevCartList);
-        setLocalCartList(parsedPrevCartList);
-      }
-    }
-  }, [cartList, setLocalCartList]);
 
   useEffect(() => {
     return () => {
