+++ src/hooks/useLocalCart.ts
@@ -1,8 +1,10 @@
 import { CartItem } from '@/types/cart';
-import { useCallback, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
+import { useCartsQuery } from './query';
 
 const useLocalCart = () => {
   const [localCartList, setLocalCartList] = useState<CartItem[]>([]);
+  const { data: cartList } = useCartsQuery();
 
   const addLocalCartItem = useCallback((cartItem: CartItem) => {
     const prevCartList = localStorage.getItem('cart');
@@ -40,6 +42,30 @@ const useLocalCart = () => {
       setLocalCartList(updatedCartList);
     }
   }, []);
+
+  useEffect(() => {
+    if (cartList?.length) {
+      const prevLocalCartList = localStorage.getItem('cart');
+
+      const parsedLocalCartList = prevLocalCartList ? JSON.parse(prevLocalCartList) : [];
+
+      const filteredCartList = cartList.filter(
+        (cartItem) =>
+          !parsedLocalCartList.find((localCartItem: CartItem) => localCartItem.productId === cartItem.productId)
+      );
+
+      const newCartList = [...parsedLocalCartList, ...filteredCartList];
+
+      localStorage.setItem('cart', JSON.stringify(newCartList));
+      setLocalCartList(newCartList);
+    } else {
+      const prevCartList = localStorage.getItem('cart');
+      if (prevCartList) {
+        const parsedPrevCartList = JSON.parse(prevCartList);
+        setLocalCartList(parsedPrevCartList);
+      }
+    }
+  }, [cartList]);
 
   return { localCartList, setLocalCartList, addLocalCartItem, deleteLocalCartItem, updateLocalCartItem };
 };
